test: keep unit tests in the default tier; add bench-integration
Address review: several files were tagged //go:build integration wholesale even
though they held pure-unit tests, excluding those from the default `go test
./...` run.

- Split data/postgres/{lifecycle,migrator,tx}_test.go and
  data/migration/{migration,group}_test.go: pure-unit tests (option/close/nil
  and no-live-connection cases) stay untagged; the DB-backed tests move to new
  *_integration_test.go files, matching data/redis and data/mongo.
- data/opensearch/lifecycle_test.go: drop the leftover FORGE_TEST_OPENSEARCH_ADDR
  branch from the untagged unit test so the default tier never touches the
  network; the live path already lives in lifecycle_integration_test.go.
- justfile: add a bench-integration recipe so the //go:build integration
  benchmarks in async/queue/{postgres,redis} stay reachable.
