+++ data/postgres/lifecycle_test.go
@@ -1,18 +1,14 @@
-//go:build integration
-
 package postgres_test
 
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/dmitrymomot/forge/data/postgres"
-	"github.com/dmitrymomot/forge/testkit/pgtest"
 )
 
 func TestClose_NilLoggerTolerated(t *testing.T) {
@@ -27,24 +23,4 @@ func TestClose_NilLoggerTolerated(t *testing.T) {
 	pool, err := pgxpool.NewWithConfig(context.Background(), cfg)
 	require.NoError(t, err)
 	assert.NotPanics(t, func() { postgres.Close(pool, nil) })
-}
-
-func TestHealthcheck_Integration(t *testing.T) {
-	dsn := pgtest.DSN(t)
-	cfg := postgres.DefaultConfig()
-	cfg.URL = dsn
-	pool, err := postgres.Open(context.Background(), postgres.WithConfig(cfg))
-	require.NoError(t, err)
-	defer postgres.Close(pool, nil)
-
-	check := postgres.Healthcheck(pool)
-	require.NotNil(t, check)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	require.NoError(t, check(ctx), "a live pool must report healthy")
-
-	// After Close, the same closure must report an ErrHealthcheck-wrapped failure.
-	postgres.Close(pool, nil)
-	assert.ErrorIs(t, check(ctx), postgres.ErrHealthcheck)
 }
