+++ data/postgres/migrator_test.go
@@ -1,11 +1,8 @@
-//go:build integration
-
 package postgres_test
 
 import (
 	"context"
 	"database/sql"
-	"errors"
 	"testing"
 	"time"
 
@@ -13,10 +10,10 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/dmitrymomot/forge/data/postgres"
-	"github.com/dmitrymomot/forge/testkit/pgtest"
 )
 
-// fakeMigrator records whether Up was called and returns a fixed error.
+// fakeMigrator records whether Up was called and returns a fixed error. It is
+// shared with the integration tier (migrator_integration_test.go).
 type fakeMigrator struct {
 	called bool
 	err    error
@@ -54,33 +51,4 @@ func TestWithMigrator_NotRunWhenConnectFails(t *testing.T) {
 	assert.ErrorIs(t, err, postgres.ErrConnect)
 	assert.Nil(t, pool)
 	assert.False(t, fm.called, "migrator must not run when the pool never came up")
-}
-
-func TestWithMigrator_RunsAfterConnect_Integration(t *testing.T) {
-	dsn := pgtest.DSN(t)
-	cfg := postgres.DefaultConfig()
-	cfg.URL = dsn
-
-	// Success path: the migrator runs and Open returns a live pool.
-	ok := &fakeMigrator{}
-	pool, err := postgres.Open(context.Background(),
-		postgres.WithConfig(cfg),
-		postgres.WithMigrator(ok),
-	)
-	require.NoError(t, err)
-	require.NotNil(t, pool)
-	assert.True(t, ok.called, "migrator must run after a live+pinged pool")
-	postgres.Close(pool, nil)
-
-	// Failure path: a migrator error fails Open and leaks no pool.
-	boom := errors.New("migration boom")
-	fail := &fakeMigrator{err: boom}
-	pool, err = postgres.Open(context.Background(),
-		postgres.WithConfig(cfg),
-		postgres.WithMigrator(fail),
-	)
-	require.Error(t, err)
-	assert.ErrorIs(t, err, boom, "a failed migration must surface as a failed Open")
-	assert.Nil(t, pool)
-	assert.True(t, fail.called)
 }
